Update level 1 & 2 to adjust the difficulty curve
Switch maps to start with a small, easy to traverse level.
Some level 1 enemies are dangerous because they will cross the player's path  but most enemies never reach the player and kill each other (see bottom right Sertails). They are just there to add pressure to the player and make them feel overwhelmed.
Level 1 was also re-centered with addded padding on the left.
Level 2 takes the whole window and has a lot of enemies (unchanged).
Level 3 is untouched for now (still no enemies) but it had unused § characters removed.
Player start position and the stairs were also updated to match level order (player starts at the right of the screen on level 1).